+++ layout/includes/accessibilitypages.php
@@ -30,21 +30,37 @@ require_once($CFG->dirroot.'/theme/boost_union/locallib.php');
 
 $config = get_config('theme_boost_union');
 
-// The static pages to be supported.
-$pages = ['accessibility', 'accessibilitysupport'];
+// The accessibility pages to be supported.
+$pages = ['declaration', 'support'];
 
-// Initialize flags if any static page should be linked from the footer and footnote.
+// The accessibility pages to be supported for logged-in users only.
+$loggedinpages = ['support'];
+
+// Initialize flags if any accessibility page should be linked from the footer.
 $templatecontext['anyaccessibilitypagelinkedfromfooter'] = false;
-$templatecontext['anyaccessibilitypagelinkedfromfootnote'] = false;
+// For the footnote, we continue to use the 'anystaticpagelinkedfromfootnote' flag from the static pages.
 
-// Iterate over the static pages.
+// Iterate over the accessibility pages.
 foreach ($pages as $page) {
     // If the page is enabled.
-    if (isset($config->{'enable'.$page}) &&
-            $config->{'enable'.$page} == THEME_BOOST_UNION_SETTING_SELECT_YES) {
+    if (isset($config->{'enableaccessibility'.$page}) &&
+            $config->{'enableaccessibility'.$page} == THEME_BOOST_UNION_SETTING_SELECT_YES) {
+
+        // If the page is only for logged-in users.
+        if (in_array($page, $loggedinpages)) {
+
+            // If the page is configured to be viewed by logged-in users only and if the user is not logged in or a guest user.
+            $allowwithoutlogin = get_config('theme_boost_union', 'allowaccessibility'.$page.'withoutlogin');
+            if (isset($allowwithoutlogin) && $allowwithoutlogin == THEME_BOOST_UNION_SETTING_SELECT_NO &&
+                    (!isloggedin() || isguestuser())) {
+                // Skip this page.
+                continue;
+            }
+        }
+
         // If the admin wants to show a link in the footnote or in both locations.
-        if ($config->{$page.'linkposition'} == THEME_BOOST_UNION_SETTING_STATICPAGELINKPOSITION_FOOTNOTE ||
-                $config->{$page.'linkposition'} == THEME_BOOST_UNION_SETTING_STATICPAGELINKPOSITION_BOTH) {
+        if ($config->{'accessibility'.$page.'linkposition'} == THEME_BOOST_UNION_SETTING_STATICPAGELINKPOSITION_FOOTNOTE ||
+                $config->{'accessibility'.$page.'linkposition'} == THEME_BOOST_UNION_SETTING_STATICPAGELINKPOSITION_BOTH) {
             // If the footnote is empty and not configured to be shown yet.
             if (isset($templatecontext['showfootnote']) == false || $templatecontext['showfootnote'] == false) {
                 // Add marker to show the footnote to templatecontext.
@@ -52,24 +68,24 @@ foreach ($pages as $page) {
             }
 
             // Add marker to show the page link in the footnote to templatecontext.
-            $templatecontext[$page.'linkpositionfootnote'] = true;
+            $templatecontext['accessibility'.$page.'linkpositionfootnote'] = true;
 
             // Flip flag that at least one page should be linked from the footnote.
             $templatecontext['anystaticpagelinkedfromfootnote'] = true;
         }
 
         // If the admin wants to show a link in the footer or in both locations.
-        if ($config->{$page.'linkposition'} == THEME_BOOST_UNION_SETTING_STATICPAGELINKPOSITION_FOOTER ||
-                $config->{$page.'linkposition'} == THEME_BOOST_UNION_SETTING_STATICPAGELINKPOSITION_BOTH) {
+        if ($config->{'accessibility'.$page.'linkposition'} == THEME_BOOST_UNION_SETTING_STATICPAGELINKPOSITION_FOOTER ||
+                $config->{'accessibility'.$page.'linkposition'} == THEME_BOOST_UNION_SETTING_STATICPAGELINKPOSITION_BOTH) {
             // Add marker to show the page link in the footer to templatecontext.
-            $templatecontext[$page.'linkpositionfooter'] = true;
+            $templatecontext['accessibility'.$page.'linkpositionfooter'] = true;
 
             // Flip flag that at least one page should be linked from the footer.
             $templatecontext['anyaccessibilitypagelinkedfromfooter'] = true;
         }
 
         // Add the page link and page title to the templatecontext.
-        $templatecontext[$page.'link'] = theme_boost_union_get_staticpage_link($page);
-        $templatecontext[$page.'pagetitle'] = theme_boost_union_get_staticpage_pagetitle($page);
+        $templatecontext['accessibility'.$page.'link'] = theme_boost_union_get_accessibility_link($page);
+        $templatecontext['accessibility'.$page.'pagetitle'] = theme_boost_union_get_accessibility_pagetitle($page);
     }
 }